# src/app/api/api-keys/route.ts
import { NextRequest, NextResponse } from "next/server";
import { createClient } from "@/lib/supabase/server";
import { authenticateWithToken } from "@/lib/supabase/service";
import { createApiKeySchema } from "@/lib/validations";
import { generateApiKey, hashApiKey, getKeyPrefix } from "@/lib/api-keys";
import { checkRateLimit, rateLimitExceeded, addRateLimitHeaders, getRateLimitIdentifier } from "@/lib/rate-limit";
import type { SupabaseClient, User } from "@supabase/supabase-js";
import type { Database } from "@/types/database";

/**
 * Get authenticated user from either cookie session or Bearer token.
 * Returns { user, supabase } or null if not authenticated.
 */
async function getAuthenticatedUser(request: NextRequest): Promise<{
  user: User;
  supabase: SupabaseClient<Database>;
} | null> {
  // First try Bearer token auth (for programmatic access)
  const authHeader = request.headers.get("authorization");
  if (authHeader?.startsWith("Bearer ")) {
    const result = await authenticateWithToken(authHeader);
    if (result) {
      return result;
    }
  }

  // Fall back to cookie session auth
  const supabase = await createClient();
  const { data: { user }, error } = await supabase.auth.getUser();
  
  if (error || !user) {
    return null;
  }

  return { user, supabase };
}

// GET /api/api-keys - List user's API keys
export async function GET(request: NextRequest) {
  try {
    const auth = await getAuthenticatedUser(request);

    if (!auth) {
      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
    }

    const { user, supabase } = auth;

    const rl = checkRateLimit(getRateLimitIdentifier(request, user.id), "read");
    if (!rl.allowed) return rateLimitExceeded(rl);

    const { data: keys, error } = await supabase
      .from("api_keys")
      .select("id, name, key_prefix, scope, last_used_at, expires_at, created_at, revoked_at")
      .eq("user_id", user.id)
      .is("revoked_at", null)
      .order("created_at", { ascending: false });

    if (error) {
      return NextResponse.json({ error: error.message }, { status: 400 });
    }

    return addRateLimitHeaders(NextResponse.json({ keys }), rl);
  } catch {
    return NextResponse.json(
      { error: "An unexpected error occurred" },
      { status: 500 }
    );
  }
}

// POST /api/api-keys - Create a new API key
export async function POST(request: NextRequest) {
  try {
    const auth = await getAuthenticatedUser(request);

    if (!auth) {
      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
    }

    const { user, supabase } = auth;

    const rl = checkRateLimit(getRateLimitIdentifier(request, user.id), "write");
    if (!rl.allowed) return rateLimitExceeded(rl);

    const body = await request.json();
    const validationResult = createApiKeySchema.safeParse(body);

    if (!validationResult.success) {
      return NextResponse.json(
        { error: validationResult.error.issues[0].message },
        { status: 400 }
      );
    }

    const { name, expires_at, scope } = validationResult.data;

    // Check if user already has a key with this name
    const { data: existingKey } = await supabase
      .from("api_keys")
      .select("id")
      .eq("user_id", user.id)
      .eq("name", name)
      .is("revoked_at", null)
      .maybeSingle();

    if (existingKey) {
      return NextResponse.json(
        { error: "You already have an active API key with this name" },
        { status: 400 }
      );
    }

    // Generate and hash the key
    const rawKey = generateApiKey(scope);
    const keyHash = await hashApiKey(rawKey);
    const keyPrefix = getKeyPrefix(rawKey);

    const { data: apiKey, error } = await supabase
      .from("api_keys")
      .insert({
        user_id: user.id,
        name,
        key_hash: keyHash,
        key_prefix: keyPrefix,
        expires_at: expires_at || null,
        scope,
      })
      .select("id, name, key_prefix, created_at, expires_at, scope")
      .single();

    if (error || !apiKey) {
      console.error("API key creation error:", error);
      return NextResponse.json({ error: error?.message || "Failed to create API key" }, { status: 400 });
    }

    // Return the full key only once at creation
    return NextResponse.json(
      {
        ...apiKey,
        key: rawKey,
      },
      { status: 201 }
    );
  } catch {
    return NextResponse.json(
      { error: "An unexpected error occurred" },
      { status: 500 }
    );
  }
}