# src/app/api/applications/route.ts
import { NextRequest, NextResponse } from "next/server";
import { applicationSchema } from "@/lib/validations";
import { sendEmail, newApplicationEmail } from "@/lib/email";
import { getAuthContext, createServiceClient } from "@/lib/auth/get-user";
import { checkRateLimit, rateLimitExceeded, getRateLimitIdentifier } from "@/lib/rate-limit";
import { dispatchWebhookAsync } from "@/lib/webhooks/dispatch";
import { getUserDid, onApplicationSubmitted } from "@/lib/reputation-hooks";
import { logActivity } from "@/lib/activity";

// POST /api/applications - Submit an application
export async function POST(request: NextRequest) {
  try {
    const auth = await getAuthContext(request);
    if (!auth) {
      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
    }
    const { user, supabase } = auth;

    const rl = checkRateLimit(getRateLimitIdentifier(request, user.id), "write");
    if (!rl.allowed) return rateLimitExceeded(rl);

    let body: unknown;
    try {
      const text = await request.text();
      if (!text || text.trim().length === 0) {
        return NextResponse.json({ error: "Request body is required" }, { status: 400 });
      }
      body = JSON.parse(text);
    } catch {
      return NextResponse.json({ error: "Invalid JSON body" }, { status: 400 });
    }

    const validationResult = applicationSchema.safeParse(body);

    if (!validationResult.success) {
      return NextResponse.json(
        { error: validationResult.error.issues[0].message },
        { status: 400 }
      );
    }

    const { gig_id, ...applicationData } = validationResult.data;

    // Check if gig exists and is active, fetch poster info for email
    const { data: gig } = await supabase
      .from("gigs")
      .select("poster_id, status, title, poster:profiles!poster_id(full_name, username)")
      .eq("id", gig_id)
      .single();

    if (!gig) {
      return NextResponse.json({ error: "Gig not found" }, { status: 404 });
    }

    // Normalize poster data
    const poster = Array.isArray(gig.poster) ? gig.poster[0] : gig.poster;

    if (gig.status !== "active") {
      return NextResponse.json(
        { error: "This gig is no longer accepting applications" },
        { status: 400 }
      );
    }

    // Can't apply to own gig
    if (gig.poster_id === user.id) {
      return NextResponse.json(
        { error: "You cannot apply to your own gig" },
        { status: 400 }
      );
    }

    // Check if already applied
    const { data: existingApplication } = await supabase
      .from("applications")
      .select("id")
      .eq("gig_id", gig_id)
      .eq("applicant_id", user.id)
      .single();

    if (existingApplication) {
      return NextResponse.json(
        { error: "You have already applied to this gig" },
        { status: 400 }
      );
    }

    // Create application
    const { data: application, error } = await supabase
      .from("applications")
      .insert({
        gig_id,
        applicant_id: user.id,
        ...applicationData,
      })
      .select()
      .single();

    if (error) {
      return NextResponse.json({ error: error.message }, { status: 400 });
    }

    // Track reputation
    const userDid = await getUserDid(supabase, user.id);
    if (userDid) {
      onApplicationSubmitted(userDid, gig_id);
    }

    // Note: notification is created by DB trigger (notify_on_new_application)
    // Do NOT insert a duplicate notification here.

    // Send email notification to gig poster
    // Get poster email from auth.users (not in profiles table)
    const adminClient = createServiceClient();
    const { data: posterAuth } = await adminClient.auth.admin.getUserById(gig.poster_id);
    const posterEmail = posterAuth?.user?.email;

    if (posterEmail) {
      // Get applicant profile
      const { data: applicantProfile } = await supabase
        .from("profiles")
        .select("full_name, username")
        .eq("id", user.id)
        .single();

      const applicantName = applicantProfile?.full_name || applicantProfile?.username || "A candidate";
      const posterName = poster?.full_name || poster?.username || "there";

      const emailContent = newApplicationEmail({
        posterName,
        applicantName,
        gigTitle: gig.title,
        gigId: gig_id,
        applicationId: application.id,
        coverLetterPreview: applicationData.cover_letter,
      });

      await sendEmail({
        to: posterEmail,
        ...emailContent,
      });
    }

    // Log activity
    void logActivity(supabase, {
      userId: user.id,
      activityType: "gig_applied",
      referenceId: gig_id,
      referenceType: "gig",
      metadata: { gig_title: gig.title },
    });

    // Dispatch webhook to gig poster
    dispatchWebhookAsync(gig.poster_id, "application.new", {
      application_id: application.id,
      gig_id,
      gig_title: gig.title,
      applicant_id: user.id,
    });

    return NextResponse.json({ application }, { status: 201 });
  } catch {
    return NextResponse.json(
      { error: "An unexpected error occurred" },
      { status: 500 }
    );
  }
}